add check producer and evaluator
